Breaking change: throw on malformed torrent
This is a breaking change.

Throwing when there's a malformed torrent allows the user of this
module to glean more information about why the torrent could not be
parsed, because the exception can be caught and contains a useful error
message. Previously, a `null` would be returned with little information.