fix: Greatly improve efficiency with torrents with a large number of files (#198)
Update the file offset algorithm from O(n*log(n)) to O(n). When generating
the files property the offset for each file was generated by summing the
lengths of all the files before. Instead cache the sum of the previous
lengths. This is a very noticeable speed up for torrents of 100,000 files
or more.